% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abx_idx_df}
\alias{abx_idx_df}
\title{A dataframe of bacterial taxonomy ranks and their cell wall, oxygen tolerance, and antibiotics characteristics}
\format{
A \code{data.frame} with columns which are:

\describe{
\item{attribute}{A characteristic of the bacteria taxonomic rank}
\item{boo}{A boolean determinant about the attribute}
\item{name}{Name of the taxon}
\item{rank}{Rank of the taxon}
\item{doi}{Reference of the attribute}
}
}
\usage{
abx_idx_df
}
\description{
A dataframe of taxonomic ranks, their phenotypes, and their antibiotics indices with references
}
\keyword{datasets}
