% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{abx_idx_plot}
\alias{abx_idx_plot}
\title{Function to plot antibiotics indices for each sample}
\usage{
abx_idx_plot(abx_vector, order = F)
}
\arguments{
\item{abx_vector}{The antibiotics vector generated from any of the abxidx functions}

\item{order}{Order the samples from highest to lowest antibiotics index before plotting}
}
\value{
A plot for the antibiotics index for each sample
}
\description{
Function to plot antibiotics indices for each sample
}
\examples{
vanco_idx <- apply(abx_test_df, 2, vancomycin_index, row.names(abx_test_df))
abx_idx_plot(vanco_idx)

}
