% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{aerobe_list}
\alias{aerobe_list}
\title{Function to return susceptibility or resistance to antibiotics targeting aerobes}
\usage{
aerobe_list(abundance, lineage)
}
\arguments{
\item{abundance}{A vector of relative abundances of bacterial taxons for a single sample}

\item{lineage}{Name of taxonomy lineage for each relative abundance in a sample (e.g. k__Bacteria; p__Bacteroidetes; c__Bacteroidia etc.)}
}
\value{
A dataframe of taxa abundances and whether they are susceptible or resistant to anti-aerobic antibiotics
}
\description{
Function to return susceptibility or resistance to antibiotics targeting aerobes
}
\examples{
do.call(rbind, apply(abx_test_df, 2, aerobe_list, row.names(abx_test_df)))

}
