% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{aminoglycoside_index}
\alias{aminoglycoside_index}
\title{Calculate the antibiotic index for aminoglycoside antibiotics}
\usage{
aminoglycoside_index(
  abundance,
  lineage,
  antibiotic_db = taxon_susceptibility,
  phenotype_db = taxon_phenotypes
)
}
\arguments{
\item{abundance}{A vector of taxon abundances in a sample}

\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{antibiotic_db}{A data frame with columns named "taxon", "rank",
"antibiotic", and "value"}

\item{phenotype_db}{A data frame with columns named "taxon", "rank", and
"gram_stain"}
}
\value{
The aminoglycoside antibiotic index for the sample
}
\description{
Calculate the antibiotic index for aminoglycoside antibiotics
}
\examples{
h22 <- weiss2021_data[weiss2021_data$sample_id \%in\% "Healthy.22",]
aminoglycoside_index(h22$proportion, h22$lineage)
}
