% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptibility.R
\name{aminoglycoside_susceptibility}
\alias{aminoglycoside_susceptibility}
\title{Determine susceptibility to aminoglycoside antibiotics}
\usage{
aminoglycoside_susceptibility(
  lineage,
  antibiotic_db = taxon_susceptibility,
  phenotype_db = taxon_phenotypes
)
}
\arguments{
\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{antibiotic_db}{A data frame with columns named "taxon", "rank",
"antibiotic", and "value"}

\item{phenotype_db}{A data frame with columns named "taxon", "rank",
"gram_stain", and "aerobic_status"}
}
\value{
A vector of assigned susceptibility values, which should be either
"susceptible", "resistant", or \code{NA}
}
\description{
Determine susceptibility to aminoglycoside antibiotics
}
