% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{anaerobe_list}
\alias{anaerobe_list}
\title{Function to return susceptibility or resistance to antibiotics targeting anaerobes}
\usage{
anaerobe_list(abundance, lineage)
}
\arguments{
\item{abundance}{A vector of relative abundances of bacterial taxons for a single sample}

\item{lineage}{Name of taxonomy lineage for each relative abundance in a sample (e.g. k__Bacteria; p__Bacteroidetes; c__Bacteroidia etc.)}
}
\value{
A dataframe of taxa abundances and whether they are susceptible or resistant to anti-anaerobic antibiotics
}
\description{
Function to return susceptibility or resistance to antibiotics targeting anaerobes
}
\examples{
do.call(rbind, apply(abx_test_df, 2, anaerobe_list, row.names(abx_test_df)))

}
