% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptibility.R
\name{antibiotic_specific_susceptibility}
\alias{antibiotic_specific_susceptibility}
\alias{antibiotic_susceptibility_vancomycin}
\alias{antibiotic_susceptibility_oxacillin}
\alias{antibiotic_susceptibility_tetracycline}
\alias{antibiotic_susceptibility_penicillin}
\alias{antibiotic_susceptibility_aminoglycoside}
\title{Determine susceptibility to vancomycin}
\usage{
antibiotic_susceptibility_vancomycin(
  lineage,
  antibiotic_db = taxon_susceptibility,
  phenotype_db = taxon_phenotypes
)

antibiotic_susceptibility_oxacillin(
  lineage,
  antibiotic_db = taxon_susceptibility,
  phenotype_db = taxon_phenotypes
)

antibiotic_susceptibility_tetracycline(
  lineage,
  antibiotic_db = taxon_susceptibility
)

antibiotic_susceptibility_penicillin(
  lineage,
  antibiotic_db = taxon_susceptibility
)

antibiotic_susceptibility_aminoglycoside(
  lineage,
  antibiotic_db = taxon_susceptibility,
  phenotype_db = taxon_phenotypes
)
}
\arguments{
\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{antibiotic_db}{A data frame with columns named "taxon", "rank",
"antibiotic", and "value"}

\item{phenotype_db}{A data frame with columns named "taxon", "rank", and
"gram_stain"}
}
\value{
A vector of assigned susceptibility values, which should be either
"susceptible", "resistant", or \code{NA}
}
\description{
Determine susceptibility to vancomycin
}
