% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{get_suscept_vector}
\alias{get_suscept_vector}
\title{Calculate the susceptiblity vector from taxonomic levels from the split_levels function}
\usage{
get_suscept_vector(row_names, abx)
}
\arguments{
\item{row_names}{The row name of taxonomic levels from the abundance matrix}

\item{abx}{What antibiotics to get the susceptibility for}
}
\value{
A vector of numbers for each taxon between 0 and 1, where 0 is resistant and 1 is susceptible. These numbers are calculated as what fractions of the species within the LTP database that are susceptible
}
\description{
Calculate the susceptiblity vector from taxonomic levels from the split_levels function
}
