% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{gram_neg_list}
\alias{gram_neg_list}
\title{Function to return susceptibility or resistance to antibiotics targeting gram negative bacteria}
\usage{
gram_neg_list(abundance, lineage)
}
\arguments{
\item{abundance}{A vector of relative abundances of bacterial taxons for a single sample}

\item{lineage}{Name of taxonomy lineage for each relative abundance in a sample (e.g. k__Bacteria; p__Bacteroidetes; c__Bacteroidia etc.)}
}
\value{
A dataframe of taxa abundances and whether they are susceptible or resistant to anti-gram negative antibiotics
}
\description{
Function to return susceptibility or resistance to antibiotics targeting gram negative bacteria
}
\examples{
do.call(rbind, apply(abx_test_df, 2, gram_neg_list, row.names(abx_test_df)))

}
