% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{gram_negative_index}
\alias{gram_negative_index}
\title{Calculate the antibiotic index for Gram-negative bacteria}
\usage{
gram_negative_index(abundance, lineage, phenotype_db = taxon_phenotypes)
}
\arguments{
\item{abundance}{A vector of taxon abundances in a sample}

\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{phenotype_db}{A data frame with columns named "taxon", "rank", and
"gram_stain"}
}
\value{
The Gram-negative antibiotic index for the sample
}
\description{
Antibiotics such as polymyxin and aztreonam.
}
\examples{
h22 <- weiss2021_data[weiss2021_data$sample_id \%in\% "Healthy.22",]
gram_negative_index(h22$proportion, h22$lineage)
}
