% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grep_df}
\alias{grep_df}
\title{A dataframe of bacterial taxonomy levels and their cell wall, oxygen tolerance, and antibiotics characteristics}
\format{
A \code{data.frame} with columns which are:

\describe{
\item{attribute}{A characteristic of the bacteria taxonomic level}
\item{boo}{A boolean determinant about the attribute}
\item{name}{Name of the taxon}
\item{level}{Level of the taxon}
\item{doi}{Reference of the attribute}
}
}
\usage{
grep_df
}
\description{
A dataframe of taxonomic levels, their phenotypes, and their antibiotics indices with references
}
\keyword{datasets}
