% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{is_susceptible}
\alias{is_susceptible}
\title{Calculate the susceptiblity vector from taxonomic lineage based on the given bacterial phenotype}
\usage{
is_susceptible(taxa, idx)
}
\arguments{
\item{taxa}{The lineage name of taxonomic levels}

\item{idx}{The bacterial phenotype to calculate the susceptibility}
}
\value{
A vector of 0 and 1, where 0 is resistant and 1 is susceptible
}
\description{
Calculate the susceptiblity vector from taxonomic lineage based on the given bacterial phenotype
}
