% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LTP}
\alias{LTP}
\title{Antibiotics index table from based on taxonomic levels from the Living Tree Project}
\format{
A \code{data.frame} with columns which are:

\describe{
\item{Kingdom}{}
\item{Phylum}{}
\item{Class}{}
\item{Order}{}
\item{Family}{}
\item{Genus}{}
\item{Species}{}
\item{gram_positive}{Index for gram positive bacteria}
\item{anaerobe}{Index for anaerobes}
\item{vancomycin}{Index for vancomycin susceptibility}
}
}
\usage{
LTP
}
\description{
A dataframe of taxonomic levels, their phenotypes, and their antibiotics indices with references
}
\keyword{datasets}
