% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirix-package.R
\docType{package}
\name{mirix-package}
\alias{mirix}
\alias{mirix-package}
\title{mirix: Modeling the Susceptibility of a Bacterial Community to Antibiotics}
\description{
This package calculates an index for a given bacterial community's susceptibility to the specified antibiotics.
}
\keyword{internal}
