% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{mirix}
\alias{mirix}
\title{Calculate Microbiome Response Index (MiRIx) values}
\usage{
mirix(abundance, susceptibility, replace_zero = 1e-04)
}
\arguments{
\item{abundance}{A vector of taxon abundances in a sample}

\item{susceptibility}{A character vector of taxon susceptibility, with
values that are "susceptible", "resistant", or \code{NA}}

\item{replace_zero}{Zero-replacement value. If the numerator or denominator
is smaller than this value, they will be replaced with the number here.
This number should generally be slightly less than the minimum measurement
that could be acquired with the method used. For count data, a value of
0.5 is typical. For relative abundances, a number that is slightly lower
than the lowest relative abundance will work.}
}
\value{
The MiRIx value
}
\description{
Calculate Microbiome Response Index (MiRIx) values
}
