% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{mirix_antibiotic}
\alias{mirix_antibiotic}
\alias{mirix_vancomycin}
\alias{mirix_doxycycline}
\alias{mirix_amoxicillin}
\alias{mirix_metronidazole}
\alias{mirix_ciprofloxacin}
\alias{mirix_gentamicin}
\title{Calculate antibiotic-specific MiRIx values}
\usage{
mirix_vancomycin(
  abundance,
  lineage,
  replace_zero = 1e-04,
  antibiotic_db = whatbacteria::taxon_susceptibility,
  phenotype_db = whatbacteria::taxon_phenotypes
)

mirix_doxycycline(
  abundance,
  lineage,
  replace_zero = 1e-04,
  antibiotic_db = whatbacteria::taxon_susceptibility
)

mirix_amoxicillin(
  abundance,
  lineage,
  replace_zero = 1e-04,
  antibiotic_db = whatbacteria::taxon_susceptibility
)

mirix_metronidazole(
  abundance,
  lineage,
  replace_zero = 1e-04,
  phenotype_db = whatbacteria::taxon_phenotypes
)

mirix_ciprofloxacin(
  abundance,
  lineage,
  replace_zero = 1e-04,
  phenotype_db = whatbacteria::taxon_phenotypes
)

mirix_gentamicin(
  abundance,
  lineage,
  replace_zero = 1e-04,
  antibiotic_db = whatbacteria::taxon_susceptibility,
  phenotype_db = whatbacteria::taxon_phenotypes
)
}
\arguments{
\item{abundance}{A vector of taxon abundances in a sample}

\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{replace_zero}{Zero-replacement value. If the numerator or denominator
is smaller than this value, they will be replaced with the number here.
This number should generally be slightly less than the minimum measurement
that could be acquired with the method used. For count data, a value of
0.5 is typical. For relative abundances, a number that is slightly lower
than the lowest relative abundance will work.}

\item{antibiotic_db}{A data frame with columns named "taxon", "rank",
"antibiotic", and "value"}

\item{phenotype_db}{A data frame with columns named "taxon", "rank",
"gram_stain", and "aerobic_status"}
}
\description{
Calculate antibiotic-specific MiRIx values
}
\examples{
h22 <- weiss2021_data[weiss2021_data$sample_id \%in\% "Healthy.22",]
mirix_vancomycin(h22$proportion, h22$lineage)
}
