% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{nitroimidazoles_index}
\alias{nitroimidazoles_index}
\title{Function to calculate antibiotics index for nitroimidazole}
\usage{
nitroimidazoles_index(df, delim = "; ")
}
\arguments{
\item{df}{A dataframe of samples as columns and bacterial taxons as rows. Labeling of taxon levels in rows needs to include full taxonomic classifications (e.g. k__Bacteria; p__Bacteroidetes; c__Bacteroidia etc.)}

\item{delim}{How taxonomic levels are separated in row names}
}
\value{
The calculated antibiotics index for each sample
}
\description{
Function to calculate antibiotics index for nitroimidazole
}
\examples{
\dontrun{
nitroimidazoles_index(test_df)
}

}
