% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{penicillin_index}
\alias{penicillin_index}
\title{Function to calculate antibiotics index for Penicillin-like antibiotics}
\usage{
penicillin_index(abundance, lineage)
}
\arguments{
\item{abundance}{A list of relative abundances of bacterial taxons for a single sample}

\item{lineage}{Name of taxonomy lineage for each relative abundance in a sample (e.g. k__Bacteria; p__Bacteroidetes; c__Bacteroidia etc.)}
}
\value{
The calculated antibiotics index for the sample
}
\description{
Function to calculate antibiotics index for Penicillin-like antibiotics
}
\examples{
apply(abx_test_df, 2, penicillin_index, row.names(abx_test_df))

}
