% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptibility.R
\name{penicillin_susceptibility}
\alias{penicillin_susceptibility}
\title{Determine susceptibility to penicillin-like antibiotics}
\usage{
penicillin_susceptibility(lineage, antibiotic_db = taxon_susceptibility)
}
\arguments{
\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{antibiotic_db}{A data frame with columns named "taxon", "rank",
"antibiotic", and "value"}
}
\value{
A vector of assigned susceptibility values, which should be either
"susceptible", "resistant", or \code{NA}
}
\description{
Determine susceptibility to penicillin-like antibiotics
}
