% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{phenotype_susceptibility}
\alias{phenotype_susceptibility}
\title{Evaluate antibiotic susceptibility based on phenotype}
\usage{
phenotype_susceptibility(
  lineage,
  phenotype,
  susceptibility,
  db = whatbacteria::taxon_phenotypes
)
}
\arguments{
\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{phenotype}{The name of the column in \code{db} that contains the
phenotype of interest}

\item{susceptibility}{A character vector specifying the susceptibility for
different values of the phenotype. The names should correspond to elements
in the column specified with \code{phenotype}. The values should be either
"susceptible" or "resistant"}

\item{db}{A data frame with columns named "taxon", "rank", and the column
name specified in \code{phenotype}}
}
\value{
A vector of assigned susceptibility values, which should be either
"susceptible", "resistant", or \code{NA}
}
\description{
This is the main engine for determining antibiotic susceptibility based on
a database of phenotypes for various taxa. To run this function, you'll need
the database as well as a vector that specifies the susceptibility for each
value of the phenotype.
}
\details{
This function operates much like \code{antibiotic_susceptibility}, except
that it pulls phenotype values from the database instead of susceptibility
information. To subsequently determine susceptibility from phenotype, this
function uses the named vector provided in the \code{susceptibility}
argument.

As a reminder, the taxonomic ranks, in order from highest to lowest, are
Kingdom, Phylum, Class, Order, Family, Genus, and Species. The ranks in the
database must be capitalized, exactly as they are written here.
}
\examples{
phenotype_susceptibility(
  c("Bacteroidetes", "Firmicutes", "Firmicutes; Negativicutes"),
  "gram_stain",
  c("Gram-positive" = "susceptible", "Gram-negative" = "resistant"))
}
