% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{polymyxin_n_aztreonam_index}
\alias{polymyxin_n_aztreonam_index}
\title{Function to calculate antibiotics index for polymyxin and aztreonam}
\usage{
polymyxin_n_aztreonam_index(df, delim = "; ", weighted = T)
}
\arguments{
\item{df}{A dataframe of samples as columns and bacterial taxons as rows. Labeling of taxon levels in rows needs to include full taxonomic classifications (e.g. k__Bacteria; p__Bacteroidetes; c__Bacteroidia etc.)}

\item{delim}{How taxonomic levels are separated in row names}

\item{weighted}{Calculate index based on weighted representation in LTP database}
}
\value{
The calculated antibiotics index for each sample
}
\description{
Function to calculate antibiotics index for polymyxin and aztreonam
}
\examples{
\dontrun{
polymyxin_n_aztreonam_index(test_df)
}

}
