% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_abundance}
\alias{predict_abundance}
\title{Predict taxon abundances at given values of an index}
\usage{
predict_abundance(index_value, abundance, susceptibility)
}
\arguments{
\item{index_value}{Value or values of the index at which to make predictions.}

\item{susceptibility}{A character vector of antibiotic susceptibility, with
values that are "susceptible", "resistant", or \code{NA}.}

\item{abundances}{A vector of taxon abundances in a sample.}
}
\value{
A new vector of abundances if \code{index_value} has length 1. If
\code{index_value} is a vector longer than 1, this function returns a
matrix, where each column contains the predicted abundances.
}
\description{
Predict taxon abundances at given values of an index
}
