% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{split_levels}
\alias{split_levels}
\title{Split the taxonomic classifications in the row names into a dataframe}
\usage{
split_levels(row_names, delim = delim)
}
\arguments{
\item{row_names}{The row name of taxonomic levels from the abundance matrix}

\item{delim}{What characters each taxonomic level are separated by}
}
\value{
A dataframe of taxonomic levels
}
\description{
Split the taxonomic classifications in the row names into a dataframe
}
