% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon_phenotypes}
\alias{taxon_phenotypes}
\title{Gram stain and aerobic status of bacterial taxa}
\format{
A data frame with the following columns:
\describe{
\item{taxon}{The name of the taxon}
\item{rank}{The rank of the taxon}
\item{aerobic_status}{
The aerobic status. One of "aerobe", "facultative anaerobe", or
"obligate anaerobe".}
\item{gram_stain}{
How the taxon appears when Gram-stained. One of "Gram-positive" or
"Gram-negative".}
\item{doi}{DOI of the publication from which the information was obtained.}
}
}
\usage{
taxon_phenotypes
}
\description{
Gram stain and aerobic status of bacterial taxa
}
\keyword{datasets}
