% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon_susceptibility}
\alias{taxon_susceptibility}
\title{Antibiotic susceptibility of bacterial taxa}
\format{
A data frame with the following columns:
\describe{
\item{taxon}{The name of the taxon}
\item{rank}{The rank of the taxon}
\item{antibiotic}{The antibiotic or antibiotic class}
\item{value}{
The susceptibility of the taxon to the antibiotic, one of "susceptible"
or "resistant".}
\item{doi}{DOI of the publication from which the information was obtained.}
}
}
\usage{
taxon_susceptibility
}
\description{
Antibiotic susceptibility of bacterial taxa
}
\keyword{datasets}
