% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{A test dataframe}
\format{
An object of class \code{data.frame} with 5 rows and 4 columns.
}
\usage{
test_df
}
\description{
A dataframe with samples as columns and relative abundances of bacteria as rows
}
\keyword{datasets}
