% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_indx.R
\name{abx_index}
\alias{abx_index}
\title{Function to calculate antibiotics index}
\usage{
abx_index(df, abx = NULL)
}
\arguments{
\item{df}{A dataframe of samples as columns and bacterial taxons as rows. Labeling of taxons in rows should follow Greengenes annotations and be separated by double underscores (e.g. k__Bacteria;p__Bacteroidetes)}

\item{abx}{Antibiotics of interest to calculate index}
}
\value{
The calculated antibiotics index for each sample
}
\description{
Function to calculate antibiotics index
}
\examples{
\dontrun{
abx_index(test_df)
}

}
