% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{aerobes_index}
\alias{aerobes_index}
\title{Calculate the antibiotic index for aerobes and facultative anaerobes}
\usage{
aerobes_index(abundance, lineage, phenotype_db = taxon_phenotypes)
}
\arguments{
\item{abundance}{A vector of taxon abundances in a sample}

\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{phenotype_db}{A data frame with columns named "taxon", "rank", and
"aerobic_status"}
}
\value{
The aerobe antibiotic index for the sample
}
\description{
Antibiotics such as fluoroquinolones.
}
\examples{
h22 <- weiss2021_data[weiss2021_data$sample_id \%in\% "Healthy.22",]
aerobes_index(h22$proportion, h22$lineage)
}
