% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptibility.R
\name{antibiotic_specific_susceptibility}
\alias{antibiotic_specific_susceptibility}
\alias{vancomycin_susceptibility}
\alias{tetracycline_susceptibility}
\alias{penicillin_susceptibility}
\alias{aminoglycoside_susceptibility}
\title{Determine susceptibility to vancomycin}
\usage{
vancomycin_susceptibility(
  lineage,
  antibiotic_db = taxon_susceptibility,
  phenotype_db = taxon_phenotypes
)

tetracycline_susceptibility(lineage, antibiotic_db = taxon_susceptibility)

penicillin_susceptibility(lineage, antibiotic_db = taxon_susceptibility)

aminoglycoside_susceptibility(
  lineage,
  antibiotic_db = taxon_susceptibility,
  phenotype_db = taxon_phenotypes
)
}
\arguments{
\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{antibiotic_db}{A data frame with columns named "taxon", "rank",
"antibiotic", and "value"}

\item{phenotype_db}{A data frame with columns named "taxon", "rank", and
"gram_stain"}
}
\value{
A vector of assigned susceptibility values, which should be either
"susceptible", "resistant", or \code{NA}
}
\description{
Determine susceptibility to vancomycin
}
