% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{calc_index}
\alias{calc_index}
\title{Calculate the index based on the susceptible vector}
\usage{
calc_index(sample_vector, suscept_vector)
}
\arguments{
\item{sample_vector}{The vector of taxonomic abundances for a sample}

\item{suscept_vector}{The vector of how susceptible a species is to the specified antibiotics}
}
\value{
The calculated antibiotics index for a sample
}
\description{
Calculate the index based on the susceptible vector
}
