% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{penicillin_index}
\alias{penicillin_index}
\title{Calculate the antibiotic index for penicillin-like antibiotics}
\usage{
penicillin_index(abundance, lineage, antibiotic_db = taxon_susceptibility)
}
\arguments{
\item{abundance}{A vector of taxon abundances in a sample}

\item{lineage}{A character vector of taxonomic assignments or lineages}

\item{antibiotic_db}{A data frame with columns named "taxon", "rank",
"antibiotic", and "value"}
}
\value{
The penicillin antibiotic index for the sample
}
\description{
Calculate the antibiotic index for penicillin-like antibiotics
}
\examples{
h22 <- weiss2021_data[weiss2021_data$sample_id \%in\% "Healthy.22",]
penicillin_index(h22$proportion, h22$lineage)
}
