% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{vancomycin_index}
\alias{vancomycin_index}
\title{Function to calculate antibiotics index for vancomycin}
\usage{
vancomycin_index(df, delim = "; ", weighted = T, plot = F, plot_order = F)
}
\arguments{
\item{df}{A dataframe of samples as columns and bacterial taxons as rows. Labeling of taxons in rows should follow Greengenes annotations and be separated by double underscores (e.g. k__Bacteria;p__Bacteroidetes)}

\item{abx}{Antibiotics of interest to calculate index}
}
\value{
The calculated antibiotics index for each sample
}
\description{
Function to calculate antibiotics index for vancomycin
}
\examples{
\dontrun{
vancomycin_index(test_df)
}

}
