% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{vancomycin_index}
\alias{vancomycin_index}
\title{Function to calculate antibiotics index for vancomycin}
\usage{
vancomycin_index(df, delim = "; ")
}
\arguments{
\item{df}{A dataframe of samples as columns and bacterial taxons as rows. Labeling of taxon levels in rows needs to include full taxonomic classifications (e.g. k__Bacteria; p__Bacteroidetes; c__Bacteroidia etc.)}

\item{delim}{How taxonomic levels are separated in row names}
}
\value{
The calculated antibiotics index for each sample
}
\description{
Function to calculate antibiotics index for vancomycin
}
\examples{
\dontrun{
vancomycin_index(test_df)
}

}
