% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abx_index.R
\name{vancomycin_index}
\alias{vancomycin_index}
\title{Function to calculate antibiotics index for Vancomycin}
\usage{
vancomycin_index(abundance, lineage)
}
\arguments{
\item{abundance}{A vector of relative abundances of bacterial taxons for a single sample}

\item{lineage}{Name of taxonomy lineage for each relative abundance in a sample (e.g. k__Bacteria; p__Bacteroidetes; c__Bacteroidia etc.)}
}
\value{
The calculated antibiotics index for the sample
}
\description{
Function to calculate antibiotics index for Vancomycin
}
\examples{
apply(abx_test_df, 2, vancomycin_index, row.names(abx_test_df))

}
